requirements.txt
pip
certifi==2018.1.18
chardet==3.0.4
emoji==0.4.5
future==0.16.0
googlemaps==2.5.1
idna==2.6
jsonpickle==0.9.6
python-telegram-bot==10.0.1
requests==2.18.4
urllib3==1.22
